+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### :zap: Added
 
+- Support for BSD 2-clause "Simplified" licenses
 - Support for MIT licenses
 
 ## [1.0.0] - 2020-01-01